[ADD] : 기대평 등록 추가
