+++ src/pages/MovieDetail/MovieDetailContent.js
@@ -8,11 +8,12 @@ import { IoTicketOutline, IoPeopleOutline } from 'react-icons/io5';
 
 function MovieDetailContent({ movieList }) {
   const [share, setShare] = useState(false);
+  // console.log('movieList', movieList);
   return (
     <>
-      {movieList.map((item, idx) => {
+      {movieList.map(item => {
         return (
-          <div key={idx}>
+          <div key={item.id}>
             <div className={styles.container}>
               <BgImg backgroundImg={item.background_img} />
 
@@ -21,7 +22,6 @@ function MovieDetailContent({ movieList }) {
                 <p className={styles.title}>{item.title}</p>
                 <p className={styles.titleEng}>{item.eng_title}</p>
                 <div className={styles.btnUtil}>
-                  {/* 좋아요 */}
                   <button
                     className={styles.shareBtn}
                     onMouseOver={() => setShare(true)}
@@ -37,7 +37,7 @@ function MovieDetailContent({ movieList }) {
                 </div>
                 <div className={styles.info}>
                   <div className={styles.score}>
-                    <p className={styles.tit}>실관람 평점</p>
+                    <p className={styles.tit}>기대평 평점</p>
                     <div className={styles.flex}>
                       <RiCheckboxMultipleLine size='22' />
                       <p className={styles.number}>
@@ -56,7 +56,6 @@ function MovieDetailContent({ movieList }) {
                     </div>
                   </div>
                   <div className={styles.score}>
-                    {/* <p className={`${styles.tit} ${styles.flexCenter}`}> */}
                     <div className={`${styles.tit} ${styles.flexCenter}`}>
                       <div>누적관객수</div>
                       <div className={styles.toolTip}>
