+++ src/pages/MovieDetail/index.js
@@ -1,8 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import MovieDetailTab from '../../components/Moviedetail/MovieDetailTab';
+import { useParams, useLocation } from 'react-router-dom';
+
+import MovieDetailTabs from '../../components/Moviedetail/MovieDetailTabs';
 import MovieDetailContent from './MovieDetailContent';
-import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
+import MovieDetailTabContent from '../../components/Moviedetail/MovieDetailTabContent/MovieDetailTabContent';
 
 const Wrapper = styled.div`
   width: 1100px;
@@ -10,19 +12,39 @@ const Wrapper = styled.div`
 `;
 
 function MovieDetail() {
+  // const location = useLocation();
   const [movieList, setMovieList] = useState([]);
-  useEffect(() => {
+  // const params = useParams();
+  // const { movieId } = useParams();
+  // console.log('index params:', movieId);
+  // useEffect(() => {
+  //   fetch(`http://localhost:10010/movie/detail/${movieId}`)
+  //     .then(res => res.json())
+  //     .then(res => setMovieList(res.data));
+  //   console.log(setMovieList.data);
+  // }, [movieId]);
+  const movieGetLoader = () => {
     fetch(`http://localhost:10010/movie/detail/1`)
       .then(res => res.json())
       .then(res => setMovieList(res.data));
-    console.log(setMovieList.data);
+  };
+  useEffect(() => {
+    movieGetLoader();
+    // console.log(setMovieList.data);
   }, []);
+
+  // useEffect(() => {
+  //   fetch(`http://localhost:10010/movie/detail/1`)
+  //     .then(res => res.json())
+  //     .then(res => setMovieList(res.data));
+  // }, []);
   return (
     <>
       <MovieDetailContent movieList={movieList} />
-      <Wrapper>
-        <MovieDetailTab />
-        <Outlet movieList={movieList} setMovieList={setMovieList} />
+      <Wrapper movieList={movieList}>
+        <MovieDetailTabs />
+        {/* 탭 아래 내용들 주요정보 4개 */}
+        <MovieDetailTabContent movieList={movieList} />
       </Wrapper>
     </>
   );
